+++ WellnessDiary/Model/CoreData/ResearchKit/WDMCDSleepQuantityStep+CoreDataProperties.swift
@@ -17,7 +17,7 @@ extension WDMCDSleepQuantityStep {
         return NSFetchRequest<WDMCDSleepQuantityStep>(entityName: "WDMCDSleepQuantityStep")
     }
 
-    @NSManaged public var value: Float
+    @NSManaged public var value: Double
     @NSManaged public var survey: WDMCDSurvey?
 
 }
